Electronic Container - Removed non used classes: Container and Electronic order new.
