+++ models/electronic_new/__init__.py
@@ -1,14 +1,14 @@
 # -*- coding: utf-8 -*-
 
-# New 
+# Commons 
 from . import txt_line
 from . import texto
 
-#from . import container
+# Electronic
 from . import electronic_line
 from . import electronic_order
-from . import electronic_order_new
 from . import electronic_container
 
+# Account
 from . import account_line
 from . import account_contasis
